Refactor app layer and add MCP installer
